Prefab Cache check.
also moved to universal patcher as it is the same for both games.
